+++ front/src/App.js
@@ -3,6 +3,8 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { RoomSetting } from "./pages/RoomSetting";
 import { Chat } from "./pages/Chat";
 import { CompleteRoomSetting } from "./pages/CompleteRoomSetting";
+import { UserSetting } from "./pages/UserSetting";
+
 
 function App() {
   return (
@@ -11,6 +13,7 @@ function App() {
         <Routes>
           <Route path="/" element={<RoomSetting />} />
           <Route path="/complete" element={<CompleteRoomSetting />} />
+          <Route path="/join/:id" element={<UserSetting />} />
           <Route path="/chat/:id" element={<Chat />} />
         </Routes>
       </BrowserRouter>
