+++ front/src/components/atoms/MessageInput.jsx
@@ -4,7 +4,7 @@ export const MessageInput = ({ member_id, room_id, sendMessage }) => {
   const [text, setText] = useState("");
   const [size, setSize] = useState("small");
 
-  const handleSendMessage = useCallback(() => {
+  const handleSubmit = useCallback(() => {
     const json = JSON.stringify({
       message: text,
       size: size,
@@ -17,7 +17,19 @@ export const MessageInput = ({ member_id, room_id, sendMessage }) => {
   }, []);
   return (
     <div className="mb-5 w-screen">
-      <input type="text" className="w-2/3 h-10 rounded-2xl " />
+      <input
+        type="text"
+        className="w-2/3 h-10 rounded-2xl"
+        onChange={(e) => {
+          setText(e.target.value);
+        }}
+      />
+      <button
+        className="text-4xl p-3 bg-violet-400 text-white rounded-lg ml-5 h-10 text-center align-middle"
+        onClick={handleSubmit}
+      >
+        Sent
+      </button>
     </div>
   );
 };
