+++ front/src/hooks/api/useUserSetting.js
@@ -1,6 +1,8 @@
 import { $axios } from "./axios";
+import { useNavigate } from "react-router-dom";
 
 export const useUserSetting = () => {
+  const navigate = useNavigate();
   const joinRoom = async (data, id) => {
     await $axios
       .post("/member/create", {
@@ -10,6 +12,7 @@ export const useUserSetting = () => {
       .then((res) => {
         //画面遷移するならここでnavigate
         console.log(res);
+        navigate('/chat/' + id)
       })
       .catch((err) => {
         console.log(err);
